requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
gunicorn>=23.0
uvicorn>=0.30
whitenoise[brotli]>=6.7
dj-database-url>=2.3
psycopg2-binary>=2.9
django-cors-headers>=4.4
python-dotenv>=1.0
joblib>=1.3
scikit-learn>=1.7
pandas>=2.2
mysqlclient>=2.2
imbalanced-learn>=0.12

